requirements.txt
python-telegram-bot[job-queue]
httpx
python-dotenv
aiohttp
psycopg2-binary
